Prevent ECP params loss
`ecpOrigin` and `partnerId` ECP params are provided through URL query params. When using the brand new routing system,  the query param are lost after a redirection. It means, if we navigate and refresh the page, params will be lost and ECP won't load.

Solution is now to store the ECP params in session storage so that it persists during the navigation.